% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyPackage.R
\docType{package}
\name{toyPackage}
\alias{toyPackage}
\alias{toyPackage-package}
\title{A toy package}
\description{
This does nothing. I mean nothing.
What’re you looking at?
}
\details{
We can put more verbiage here if we want.
}

