% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{x}
\alias{x}
\title{A bunch of radii (TITLE)}
\format{description of the variables,
  with full names and explanations (esp
  for data frames). Dimensions help.}
\source{
where did the data come from?
  This could be a URL.
}
\description{
Really, do you need this explained?
Well, ok. (DESCRIPTION)
}

